Вносит правки по отрисовке фотографий
