+++ js/main.js
@@ -1,4 +1,5 @@
+import { renderPictures } from './pictures.js';
 import { generatePhotos } from './data.js';
 
-generatePhotos();
-
+const pictures = generatePhotos();
+renderPictures(pictures);
